メニュー関数修正
