+++ src/components/layout/Header/index.module.scss
@@ -1,4 +1,3 @@
-/* stylelint-disable rule-empty-line-before */
 @use '@/styles' as *;
 
 .header {
